Reduce complexity of Sample Receiver.
